feature/GH-5583: rename files
