+++ projects/storefrontlib/src/cms-components/myaccount/order/cancellations-returns/cancel-or-return-items/cancel-or-return-items.component.ts
@@ -12,7 +12,7 @@ import {
   CancelOrReturnRequestEntryInput,
   Price,
 } from '@spartacus/core';
-import { OrderCancelOrReturnService } from '../cancel-or-returns.service';
+import { OrderCancelOrReturnService } from '../cancel-or-return.service';
 
 @Component({
   selector: 'cx-cancel-or-return-items',
