+++ projects/storefrontlib/src/cms-components/myaccount/order/cancellations-returns/cancel-or-return.module.ts
@@ -1,7 +1,7 @@
 import { NgModule } from '@angular/core';
 import { ReturnOrderModule } from './return-order/return-order.module';
 import { ReturnOrderConfirmationModule } from './return-order-confirmation/return-order-confirmation.module';
-import { OrderCancelOrReturnService } from './cancel-or-returns.service';
+import { OrderCancelOrReturnService } from './cancel-or-return.service';
 
 @NgModule({
   imports: [ReturnOrderModule, ReturnOrderConfirmationModule],
