+++ projects/storefrontlib/src/cms-components/myaccount/order/cancellations-returns/index.ts
@@ -1,4 +1,4 @@
 export * from './cancel-or-return-items/cancel-or-return-items.module';
 export * from './return-order/return-order.module';
 export * from './return-order-confirmation/return-order-confirmation.module';
-export * from './cancel-or-returns.module';
+export * from './cancel-or-return.module';
