+++ projects/storefrontlib/src/cms-components/myaccount/order/order.module.ts
@@ -1,7 +1,7 @@
 import { NgModule } from '@angular/core';
 import { OrderHistoryModule } from './order-history/order-history.module';
 import { OrderDetailsModule } from './order-details/order-details.module';
-import { OrderCancelOrReturnModule } from './cancellations-returns/cancel-or-returns.module';
+import { OrderCancelOrReturnModule } from './cancellations-returns/cancel-or-return.module';
 
 @NgModule({
   imports: [OrderHistoryModule, OrderDetailsModule, OrderCancelOrReturnModule],
